# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /follow-ups:
    get:
      operationId: listFollowUps
      tags: [followups]
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/FollowUpStatus" }
        - name: type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/FollowUpType" }
        - name: dateFrom
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: dateTo
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowUp
      tags: [followups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /follow-ups/{id}:
    patch:
      operationId: updateFollowUp
      tags: [followups]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
    delete:
      operationId: deleteFollowUp
      tags: [followups]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /follow-ups/{id}/call:
    post:
      operationId: triggerFollowUpCall
      tags: [followups]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerCallResult"
...
components:
  schemas:
...
    FollowUpType:
      type: string
      enum: [RENEWAL_REMINDER, MONTHLY_CHECKIN, CALLBACK_REQUESTED, MANUAL, POLICY_ANNIVERSARY]
    FollowUpStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, SKIPPED, RESCHEDULED]
...
    TriggerCallResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        call_log_id: { type: ["string", "null"] }
        execution_id: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
...
    FollowUp:
      type: object
      required: [id, org_id, lead_id, type, scheduled_at, status, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: string }
        lead_name: { type: ["string", "null"] }
        policy_id: { type: ["string", "null"] }
        type: { $ref: "#/components/schemas/FollowUpType" }
        scheduled_at: { type: string, format: date-time }
        completed_at: { type: ["string", "null"], format: date-time }
        bolna_agent_id: { type: ["string", "null"] }
        call_log_id: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/FollowUpStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    FollowUpInput:
      type: object
      required: [lead_id, type, scheduled_at]
      properties:
        lead_id: { type: string }
        policy_id: { type: string }
        type: { $ref: "#/components/schemas/FollowUpType" }
        scheduled_at: { type: string, format: date-time }
        bolna_agent_id: { type: string }
        notes: { type: string }
    FollowUpUpdate:
      type: object
      properties:
        type: { $ref: "#/components/schemas/FollowUpType" }
        scheduled_at: { type: string, format: date-time }
        bolna_agent_id: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/FollowUpStatus" }
...